+++ libs/utils/source/ftrack_utils/extensions/registry.py
@@ -87,9 +87,9 @@ def register_js_files(file_list):
                     extension_name = line.split("=")[1].strip().strip('"')
                 content += line
         if not (extension_type and extension_name):
-            # Log warning if yaml file doesn't contain type key.
             logger.warning(
-                "No extension compatible .js file, missing 'EXTENSION_TYPE'."
+                "No extension compatible .js file, missing 'EXTENSION_TYPE' and"
+                " 'EXTENSION_NAME' definitions."
                 "\nFile: {}".format(_file)
             )
             continue