+++ libs/utils/release_notes.md
@@ -2,6 +2,8 @@
 
 ## upcoming
 * [new] Registry; Support for JS extensions.
+* [feat] JS RPC; Added on connected callback.
+* [feat] JS RPC; Aligned on_run_dialog_callback with DCC bootstrap implementation.
 
 
 ## v2.2.0