+++ libs/utils/release_notes.md
@@ -1,7 +1,7 @@
 # ftrack Utils library release Notes
 
 ## upcoming
-
+* [new] Registry; Support for JS extensions.
 * [feat] JS RPC; Added on connected callback.
 * [feat] JS RPC; Aligned on_run_dialog_callback with DCC bootstrap implementation.